pyproject.toml
[project]
name = "docmap"
version = "0.1.0"
description = "DocMap pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "requests>=2.32.0",
  "beautifulsoup4>=4.12.0",
  "apscheduler>=3.10.0",
  "google-cloud-bigquery>=3.30.0",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.9.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
